pocs/skill_inventory_bot/requirements.txt
streamlit>=1.31.0
anthropic>=0.40.0
python-dotenv>=1.0.0
